Ship movement left, range, speed
Implemented ship movement to the left, limited the moving rang to the screen width and adjusted the speed of movement.
Refactored the _check_events method.
Embedded another quitting function by pressing q. Still the programm crashes when closed.
Full screen mode could be embedded, but moving speed of the ship should be adapted to the screen size then. (As long as this feature is not availible, the game is better run in window mode.) --> Switched back to window mode like before.